+++ david/go.mod
@@ -1,0 +1,3 @@
+module david
+
+go 1.20
