+++ david/readme.md
@@ -1,0 +1,6 @@
+# Execution
+Install go [here](https://go.dev/doc/install)
+
+then run `go run main.go <day#>`
+
+eg `go run main.go 1` for day 1 etc.
